# ayunis-core-frontend/src/shared/api/openapi-schema.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/threads": {
      "post": {
        "operationId": "ThreadsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateThreadDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The thread has been successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetThreadResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid model data"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Create a new thread",
        "tags": [
          "threads"
        ]
      },
      "get": {
        "operationId": "ThreadsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns all threads for the current user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetThreadsResponseDtoItem"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Get all threads for the current user",
        "tags": [
          "threads"
        ]
      }
    },
    "/threads/{id}": {
      "get": {
        "operationId": "ThreadsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The UUID of the thread to retrieve",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the thread with the specified ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetThreadResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Thread not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Get a thread by ID",
        "tags": [
          "threads"
        ]
      },
      "delete": {
        "operationId": "ThreadsController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The UUID of the thread to delete",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The thread has been successfully deleted"
          },
          "404": {
            "description": "Thread not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Delete a thread",
        "tags": [
          "threads"
        ]
      }
    },
    "/threads/{id}/instruction": {
      "patch": {
        "operationId": "ThreadsController_updateInstruction",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The UUID of the thread to update",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateThreadInstructionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The thread instruction has been successfully updated"
          },
          "400": {
            "description": "Invalid instruction data"
          },
          "404": {
            "description": "Thread not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Update thread instruction",
        "tags": [
          "threads"
        ]
      }
    },
    "/threads/{id}/model": {
      "patch": {
        "operationId": "ThreadsController_updateModel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The UUID of the thread to update",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateThreadModelDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The thread model has been successfully updated"
          },
          "400": {
            "description": "Invalid model data"
          },
          "404": {
            "description": "Thread not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Update thread model",
        "tags": [
          "threads"
        ]
      }
    },
    "/threads/{id}/internet-search": {
      "patch": {
        "operationId": "ThreadsController_updateInternetSearch",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The UUID of the thread to update",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateThreadInternetSearchDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The thread internet search setting has been successfully updated"
          },
          "400": {
            "description": "Invalid internet search data"
          },
          "404": {
            "description": "Thread not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Update thread internet search setting",
        "tags": [
          "threads"
        ]
      }
    },
    "/threads/{id}/sources": {
      "get": {
        "operationId": "ThreadsController_getThreadSources",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The UUID of the thread",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns all sources for the thread",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SourceResponseDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Thread not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Get all sources for a thread",
        "tags": [
          "threads"
        ]
      }
    },
    "/threads/{id}/sources/file": {
      "post": {
        "operationId": "ThreadsController_addFileSource",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The UUID of the thread",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "The file to upload"
                  },
                  "userId": {
                    "type": "string",
                    "format": "uuid",
                    "description": "The ID of the user who owns this source"
                  },
                  "name": {
                    "type": "string",
                    "description": "The display name for the file source"
                  },
                  "description": {
                    "type": "string",
                    "description": "A description of the file source"
                  }
                },
                "required": [
                  "file",
                  "userId"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The file source has been successfully added to the thread",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileSourceResponseDto"
                }
              }
            }
          }
        },
        "summary": "Add a file source to a thread",
        "tags": [
          "threads"
        ]
      }
    },
    "/threads/{id}/sources/{sourceId}": {
      "delete": {
        "operationId": "ThreadsController_removeSource",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The UUID of the thread",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "sourceId",
            "required": true,
            "in": "path",
            "description": "The UUID of the source to remove",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The source has been successfully removed from the thread"
          },
          "404": {
            "description": "Thread or source not found"
          }
        },
        "summary": "Remove a source from a thread",
        "tags": [
          "threads"
        ]
      }
    },
...
  },
  "info": {
    "title": "Ayunis Core API",
    "description": "The Ayunis Core API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateThreadDto": {
        "type": "object",
        "properties": {
          "modelId": {
            "type": "string",
            "description": "The id of the model",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "format": "uuid"
          }
        },
        "required": [
          "modelId"
        ]
      },
...
      "GetThreadResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the thread",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "userId": {
            "type": "string",
            "description": "User ID who owns this thread",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "model": {
            "description": "Model",
            "allOf": [
              {
                "$ref": "#/components/schemas/ModelResponseDto"
              }
            ]
          },
          "title": {
            "type": "string",
            "description": "Title of the thread",
            "example": "Discussion about AI models"
          },
          "instruction": {
            "type": "string",
            "description": "Instruction for the thread",
            "example": "Please analyze the attached documents and provide insights on market trends."
          },
          "isInternetSearchEnabled": {
            "type": "boolean",
            "description": "Whether internet search is enabled for the thread",
            "example": false
          },
          "messages": {
            "type": "array",
            "description": "Array of messages in the thread (role-specific types)",
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/UserMessageResponseDto"
                },
                {
                  "$ref": "#/components/schemas/SystemMessageResponseDto"
                },
                {
                  "$ref": "#/components/schemas/AssistantMessageResponseDto"
                },
                {
                  "$ref": "#/components/schemas/ToolResultMessageResponseDto"
                }
              ]
            }
          },
          "createdAt": {
            "type": "string",
            "description": "Creation timestamp",
            "example": "2024-01-01T12:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "description": "Last update timestamp",
            "example": "2024-01-01T12:30:00.000Z"
          }
        },
        "required": [
          "id",
          "userId",
          "model",
          "isInternetSearchEnabled",
          "messages",
          "createdAt",
          "updatedAt"
        ]
      },
      "GetThreadsResponseDtoItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the thread",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "title": {
            "type": "string",
            "description": "Title of the thread",
            "example": "Discussion about AI models"
          },
          "createdAt": {
            "type": "string",
            "description": "Creation timestamp",
            "example": "2024-01-01T12:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "description": "Last update timestamp",
            "example": "2024-01-01T12:30:00.000Z"
          }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateThreadInstructionDto": {
        "type": "object",
        "properties": {
          "instruction": {
            "type": "string",
            "description": "The instruction for the thread",
            "example": "Please analyze the attached documents and provide insights on market trends."
          }
        },
        "required": [
          "instruction"
        ]
      },
      "UpdateThreadModelDto": {
        "type": "object",
        "properties": {
          "modelName": {
            "type": "string",
            "description": "The name of the model",
            "example": "gpt-4"
          },
          "modelProvider": {
            "type": "string",
            "description": "The provider of the model",
            "example": "openai",
            "enum": [
              "openai",
              "anthropic",
              "mistral",
              "microsoft",
              "ollama"
            ]
          }
        },
        "required": [
          "modelName",
          "modelProvider"
        ]
      },
      "UpdateThreadInternetSearchDto": {
        "type": "object",
        "properties": {
          "isInternetSearchEnabled": {
            "type": "boolean",
            "description": "Whether internet search is enabled for the thread",
            "example": true
          }
        },
        "required": [
          "isInternetSearchEnabled"
        ]
      },
      "SourceResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the source"
          },
          "threadId": {
            "type": "string",
            "description": "Thread ID this source belongs to"
          },
          "userId": {
            "type": "string",
            "description": "User ID who owns this source"
          },
          "type": {
            "type": "string",
            "description": "Type of source",
            "enum": [
              "file",
              "url"
            ]
          },
          "createdAt": {
            "type": "string",
            "description": "Creation timestamp"
          },
          "updatedAt": {
            "type": "string",
            "description": "Last update timestamp"
          }
        },
        "required": [
          "id",
          "threadId",
          "userId",
          "type",
          "createdAt",
          "updatedAt"
        ]
      },
...
    }
  }
}